+++ lib/decidim/maintainers_toolbox/github_manager/querier/by_issue_id.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "base"
+require "active_support/core_ext/object/blank"
 
 module Decidim
   module MaintainersToolbox