+++ spec/lib/decidim/maintainers_toolbox/backporter_spec.rb
@@ -2,7 +2,7 @@
 
 require "decidim/maintainers_toolbox/backporter"
 
-describe Decidim::MaintainersToolbox::Backporter do
+RSpec.describe Decidim::MaintainersToolbox::Backporter do
   subject { described_class.new(token: token, pull_request_id: pull_request_id, version_number: version_number, exit_with_unstaged_changes: exit_with_unstaged_changes, with_console: with_console) }
 
   let(:token) { "1234" }