+++ CHANGELOG.md
@@ -5,7 +5,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.3.0 - 2024-09-25
+## [0.7.0] - 2025-02-11
+
+### Changed
+
+- Change all the binstubs call to use plain old bundle exec
+
+## [0.6.0] - 2025-02-10
+
+### Changed
+
+- Make rspec call work with devenv.sh
+
+## [0.5.0] - 2025-02-10
+
+### Fixed
+
+-  Fix rspec call with environment variables
+
+## [0.4.0] - 2025-02-10
+
+### Changed
+
+- Update ruby version from 2.7.5 to 3.0.2
+
+### Fixed
+
+-  Make directory check compatible with latest versions of ruby (#11)
+
+## [0.3.0] - 2024-09-25
 
 ### Added